Move deref to a new preprocessor module
