+++ lib/preprocessor/deref.ts
