+++ lib/schema/index.ts
@@ -22,11 +22,6 @@ export {
   validateSchema
 } from './jsonschema'
 
-export {
-  JSONSchema,
-  dereferenceSchema
-} from './deref'
-
 export interface BooleanEncodingSchema {
   readonly type: 'boolean';
 }
